+++ openlane/steps/openroad.py
@@ -19,6 +19,7 @@ import re
 import subprocess
 import tempfile
 import textwrap
+import pathlib
 from abc import abstractmethod
 from base64 import b64encode
 from concurrent.futures import Future, ThreadPoolExecutor
@@ -37,6 +38,7 @@ from ..common import (
     Path,
     Filter,
     TclUtils,
+    DRC as DRCObject,
     _get_process_limit,
     aggregate_metrics,
     get_script_dir,
@@ -1168,6 +1170,13 @@ class TapEndcapInsertion(OpenROADStep):
 
     config_vars = OpenROADStep.config_vars + [
         Variable(
+            "FP_TAPCELL_DIST",
+            Optional[Decimal],
+            "The distance between tap cell columns. Must be specified if WELLTAP_CELL is specified.",
+            units="µm",
+            pdk=True,
+        ),
+        Variable(
             "FP_MACRO_HORIZONTAL_HALO",
             Decimal,
             "Specify the horizontal halo size around macros.",
@@ -1183,17 +1192,18 @@ class TapEndcapInsertion(OpenROADStep):
             units="µm",
             deprecated_names=["FP_TAP_VERTICAL_HALO"],
         ),
-        Variable(
-            "FP_TAPCELL_DIST",
-            Decimal,
-            "The distance between tap cell columns.",
-            units="µm",
-            pdk=True,
-        ),
     ]
 
     def get_script_path(self):
         return os.path.join(get_script_dir(), "openroad", "tapcell.tcl")
+
+    def run(self, state_in, **kwargs):
+        if (
+            self.config["WELLTAP_CELL"] is not None
+            and self.config["FP_TAPCELL_DIST"] is None
+        ):
+            raise StepException("FP_TAPCELL_DIST must be set if WELLTAP_CELL is set.")
+        return super().run(state_in, **kwargs)
 
 
 @Step.factory.register()
@@ -1655,6 +1665,13 @@ class RepairAntennas(CompositeStep):
 
     Steps = [_DiodeInsertion, CheckAntennas]
 
+    def run(self, state_in: State, **kwargs) -> Tuple[ViewsUpdate, MetricsUpdate]:
+        if self.config["DIODE_CELL"] is None:
+            info(f"'DIODE_CELL' not set. Skipping '{self.id}'…")
+            return {}, {}
+
+        return super().run(state_in, **kwargs)
+
 
 @Step.factory.register()
 class DetailedRouting(OpenROADStep):
@@ -1719,7 +1736,7 @@ class DetailedRouting(OpenROADStep):
         Variable(
             "DRT_SAVE_DRC_REPORT_ITERS",
             Optional[int],
-            "Report DRC on each specified iteration. Set to 1 when DRT_SAVE_DRC_REPORT_ITERS in enabled",
+            "Write a DRC report every N iterations. If DRT_SAVE_SNAPSHOTS is enabled, there is an implicit default value of 1.",
         ),
     ]
 
@@ -1730,7 +1747,21 @@ class DetailedRouting(OpenROADStep):
         kwargs, env = self.extract_env(kwargs)
         env["DRT_THREADS"] = env.get("DRT_THREADS", str(_get_process_limit()))
         info(f"Running TritonRoute with {env['DRT_THREADS']} threads…")
-        return super().run(state_in, env=env, **kwargs)
+        views_updates, metrics_updates = super().run(state_in, env=env, **kwargs)
+
+        drc_paths = list(pathlib.Path(self.step_dir).rglob("*.drc*"))
+        for path in drc_paths:
+            drc, _ = DRCObject.from_openroad(
+                open(path, encoding="utf8"), self.config["DESIGN_NAME"]
+            )
+
+            drc.to_klayout_xml(open(pathlib.Path(str(path) + ".xml"), "wb"))
+        #        if violation_count > 0:
+        #            self.warn(
+        #                f"DRC errors found after routing. View the report file at {report_path}.\nView KLayout xml file at {klayout_db_path}"
+        #            )
+
+        return views_updates, metrics_updates
 
 
 @Step.factory.register()
@@ -2187,7 +2218,7 @@ class CTS(OpenROADStep):
             Variable(
                 "CTS_CLK_BUFFERS",
                 List[str],
-                "Defines the list of clock buffers to be used in CTS.",
+                "Defines the list of clock buffer names or buffer name wildcards to be used in CTS.",
                 deprecated_names=["CTS_CLK_BUFFER_LIST"],
                 pdk=True,
             ),